requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
httpx==0.25.0
python-dotenv==1.0.0
python-multipart==0.0.6
loguru==0.7.2
ffmpeg-python==0.2.0
python-jose==3.3.0